moved every model to same folder in the root
