+++ user/backend/routes/images.js
@@ -1,8 +1,8 @@
 const express = require('express');
 const createError = require('http-errors');
 const { verifyAccessToken } = require('../../../helpers/jwt');
-const User = require('../models/User');
-const History = require('../models/History');
+const User = require('../../../models/User');
+const History = require('../../../models/History');
 const router = express.Router();
 const exec = require('child_process').exec;
 const path = require('path');
